Polished level rendering
Void tiles are now returned when the region viewed is out of bounds, not
supplied by the map.
Black borders around the edges have been repaired- Left and upper
borders, have adjusted the render cancellation to not cancel ALL of the
x values for a tile when just one is off screen >_> (upper level did not
have a break statement so Y had no box)
Right and Bottom- added one to the corner pins, to increase screen
rendering. After that it just amounts to the same as left + top