Hoo Dang. So, added the functionality to "draw" levels- that is, we can load levels from a png file.
Here is an important remark- I strayed from the video to do this. I
already reckoned fairly well what I needed to do to get the level drawn
to the screen, w/o help from the video for the most part. The nicky
thing is, I will probably have to undo a lot of what I've done to make
it work, simply to synchronize my code with TheChernos. However it
shouldnt be an issue. And this way I got to learn a lot more, by doing a
lot myself, just hopefully it is in the 'right' fashion.