Refactoring- Event handlers for the mouse
Added event handlers for the mouse, moving through a layer stack. This
way we can block mouse events from layers further in the stack if other
layers already handled them. There is a slight problem, in the
onReleased event in player, we WANT to only run it when, very
specifically, BUTTON1 is released, but as it stands we're checking for
NOBUTTON (a condition that is ALWAYS true on MouseRelease any way :I)